requirements/base.txt

# Bleeding edge Django
Django==1.9.8

# Configuration
django-environ==0.4.0
django-secure==1.0.1

# Forms
django-braces==1.8.1
django-crispy-forms==1.6.0
django-floppyforms==1.6.2

# Models
django-model-utils==2.5

# Images
Pillow==3.2.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.25.2


# Python-PostgreSQL Database Adapter
psycopg2==2.6.1

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2016.4

# Redis support
django-redis==4.4.2
redis>=2.10.0


# Your custom requirements go here

# django-cms requirements
django-cms==3.3.1
django-treebeard
djangocms-admin-style==0.2.2
django-reversion==1.10.2
django-classy-tags==0.7.2
django-sekizai==0.9.0
html5lib>=0.90,!=0.9999,!=0.99999,<0.99999999
six
django-filer
cmsplugin-filer
djangocms-text-ckeditor==3.0.1
djangocms-link
djangocms-snippet
djangocms-style
djangocms-column
djangocms-grid
djangocms-oembed
djangocms-table
djangocms-flash
djangocms-file
djangocms-googlemap
djangocms-picture
djangocms-teaser
djangocms-video
djangocms-inherit

django-pagination
django-polymorphic
django-ajax-selects
django-postman
django-notifications-hq
